+++ indexer/testing/tests/graphql/event.rs
@@ -257,3 +257,54 @@ async fn graphql_subscribe_to_events() -> anyhow::Result<()> {
         })
         .await?
 }
+
+#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+async fn graphql_returns_nested_events() -> anyhow::Result<()> {
+    let (httpd_context, _client, ..) = create_block().await?;
+
+    let graphql_query = r#"
+      query Transactions {
+        transactions {
+          nodes {
+            blockHeight
+            sender
+            hash
+            hasSucceeded
+            nestedEvents
+          }
+          edges { node { blockHeight sender hash hasSucceeded nestedEvents } cursor }
+          pageInfo { hasPreviousPage hasNextPage startCursor endCursor }
+        }
+      }
+    "#;
+
+    let request_body = GraphQLCustomRequest {
+        name: "transactions",
+        query: graphql_query,
+        variables: Default::default(),
+    };
+
+    let local_set = tokio::task::LocalSet::new();
+
+    local_set
+        .run_until(async {
+            tokio::task::spawn_local(async move {
+                let app = build_app_service(httpd_context);
+
+                let response =
+                    call_graphql::<PaginatedResponse<serde_json::Value>>(app, request_body).await?;
+
+                let nested_events: &str = response.data.edges[0]
+                    .node
+                    .get("nestedEvents")
+                    .and_then(|s| s.as_str())
+                    .expect("Can't get nestedEvents");
+
+                assert_that!(nested_events.len()).is_at_least(10);
+
+                Ok::<(), anyhow::Error>(())
+            })
+            .await
+        })
+        .await?
+}
