+++ indexer/testing/tests/graphql/main.rs
@@ -1,2 +1,5 @@
 pub mod block;
 pub mod event;
+pub mod grug;
+pub mod message;
+pub mod transaction;
