+++ indexer/testing/tests/graphql/transaction.rs
@@ -1,0 +1,222 @@
+use {
+    assert_json_diff::assert_json_include,
+    assertor::*,
+    grug_types::{BroadcastClientExt, Coins, Denom, GasOption, Message, ResultExt},
+    indexer_sql::entity,
+    indexer_testing::{
+        GraphQLCustomRequest, PaginatedResponse, block::create_block, build_app_service,
+        call_graphql, call_ws_graphql_stream, parse_graphql_subscription_response,
+    },
+    serde_json::json,
+    std::str::FromStr,
+    tokio::sync::mpsc,
+};
+
+#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+async fn graphql_returns_last_block_transactions() -> anyhow::Result<()> {
+    let (httpd_context, _client, ..) = create_block().await?;
+
+    let graphql_query = r#"
+      query Block {
+        block {
+          blockHeight
+          transactions {
+            blockHeight
+          }
+        }
+      }
+    "#;
+
+    let request_body = GraphQLCustomRequest {
+        name: "block",
+        query: graphql_query,
+        variables: Default::default(),
+    };
+
+    let local_set = tokio::task::LocalSet::new();
+
+    local_set
+        .run_until(async {
+            tokio::task::spawn_local(async {
+                let app = build_app_service(httpd_context);
+
+                let response = call_graphql::<serde_json::Value>(app, request_body).await?;
+
+                let expected = json!({
+                    "blockHeight": 1,
+                    "transactions": [
+                        {
+                            "blockHeight": 1,
+                        }
+                    ]
+                });
+
+                assert_json_include!(actual: response.data, expected: expected);
+
+                Ok::<(), anyhow::Error>(())
+            })
+            .await
+        })
+        .await?
+}
+
+#[tokio::test(flavor = "multi_thread", worker_threads = 2)]
+async fn graphql_returns_transactions() -> anyhow::Result<()> {
+    let (httpd_context, _client, accounts) = create_block().await?;
+
+    let graphql_query = r#"
+      query Transactions {
+        transactions {
+          nodes {
+            id
+            blockHeight
+            sender
+            hash
+            hasSucceeded
+            createdAt
+            transactionType
+            transactionIdx
+            data
+            credential
+            gasWanted
+            gasUsed
+            errorMessage
+          }
+          edges { node { id createdAt blockHeight sender hash hasSucceeded transactionType transactionIdx data credential gasWanted gasUsed errorMessage } cursor }
+          pageInfo { hasPreviousPage hasNextPage startCursor endCursor }
+        }
+      }
+    "#;
+
+    let request_body = GraphQLCustomRequest {
+        name: "transactions",
+        query: graphql_query,
+        variables: Default::default(),
+    };
+
+    let local_set = tokio::task::LocalSet::new();
+
+    local_set
+        .run_until(async {
+            tokio::task::spawn_local(async move {
+                let app = build_app_service(httpd_context);
+
+                let response = call_graphql::<PaginatedResponse<entity::transactions::Model>>(
+                    app,
+                    request_body,
+                )
+                .await?;
+
+                assert_that!(response.data.edges).has_length(1);
+
+                assert_that!(response.data.edges[0].node.sender)
+                    .is_equal_to(accounts["sender"].address.to_string());
+
+                Ok::<(), anyhow::Error>(())
+            })
+            .await
+        })
+        .await?
+}
+
+#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
+async fn graphql_subscribe_to_transactions() -> anyhow::Result<()> {
+    let (httpd_context, client, mut accounts) = create_block().await?;
+
+    let graphql_query = r#"
+      subscription Transactions {
+        transactions {
+          id
+          data
+          credential
+          blockHeight
+          createdAt
+          transactionType
+          transactionIdx
+          sender
+          hash
+          hasSucceeded
+          errorMessage
+          gasWanted
+          gasUsed
+        }
+      }
+    "#;
+
+    let request_body = GraphQLCustomRequest {
+        name: "transactions",
+        query: graphql_query,
+        variables: Default::default(),
+    };
+
+    let (crate_block_tx, mut rx) = mpsc::channel::<u32>(1);
+
+    // Can't call this from LocalSet so using channels instead.
+    tokio::spawn(async move {
+        while rx.recv().await.is_some() {
+            let to = accounts["owner"].address;
+            let chain_id = client.chain_id().await;
+
+            client
+                .send_message(
+                    &mut accounts["sender"],
+                    Message::transfer(to, Coins::one(Denom::from_str("ugrug")?, 2_000)?)?,
+                    GasOption::Predefined { gas_limit: 2000 },
+                    &chain_id,
+                )
+                .await
+                .should_succeed();
+
+            // Enabling this here will cause the test to hang
+            // suite.app.indexer.wait_for_finish();
+        }
+
+        Ok::<(), anyhow::Error>(())
+    });
+
+    let local_set = tokio::task::LocalSet::new();
+
+    local_set
+        .run_until(async {
+            tokio::task::spawn_local(async move {
+                let name = request_body.name;
+                let (_srv, _ws, framed) =
+                    call_ws_graphql_stream(httpd_context, build_app_service, request_body).await?;
+
+                // 1st response is always the existing last block
+                let (framed, response) = parse_graphql_subscription_response::<
+                    Vec<entity::transactions::Model>,
+                >(framed, name)
+                .await?;
+
+                assert_that!(response.data.first().unwrap().block_height).is_equal_to(1);
+                assert_that!(response.data).has_length(1);
+
+                crate_block_tx.send(2).await?;
+
+                // 2st response
+                let (framed, response) = parse_graphql_subscription_response::<
+                    Vec<entity::transactions::Model>,
+                >(framed, name)
+                .await?;
+
+                assert_that!(response.data.first().unwrap().block_height).is_equal_to(2);
+                assert_that!(response.data).has_length(1);
+
+                crate_block_tx.send(3).await?;
+
+                // 3rd response
+                let (_, response) = parse_graphql_subscription_response::<
+                    Vec<entity::transactions::Model>,
+                >(framed, name)
+                .await?;
+
+                assert_that!(response.data.first().unwrap().block_height).is_equal_to(3);
+                assert_that!(response.data).has_length(1);
+
+                Ok::<(), anyhow::Error>(())
+            })
+            .await
+        })
+        .await?
+}
